.tmx file parser
wsTotal is my complete workspace, wsTotal.zip is a zipped version if
Eclipse gives issues with missing files or parameters..
See LoadingScreen.java to see how the filename is passed and an example
of the getTile() helper function.  The .tmx file that you want to read
should be in the assets directory.  TiledMap.java is where all the
contents of the .tmx file is read into.